backend/requirements.txt
fastapi
uvicorn
supabase
stripe
httpx
pandas
openpyxl
python-dotenv
python-multipart
